Fix Loan Edit screen and resolve test failures (#44)
* Fix Loan Edit screen, update dependencies, and resolve test failures

* Fix streak calculation to include Fixed Expenses and Income

* Show streak pill in UI even if 0, update styles

* Fix SQL syntax for streak calculation to correctly order grouped results

* Add widget tests for EditLoanScreen to improve coverage

* Simplify SQL query for streak calculation and clean up tests

+++ lib/core/services/notification_service.dart
@@ -338,7 +338,7 @@ class NotificationService {
     if (!_isInitialized) await init();
 
     final int count = bills.length;
-    final int total = bills.fold(0, (sum, b) => sum + b.amount);
+    final double total = bills.fold(0.0, (sum, b) => sum + b.amount);
 
     final String title = "Daily Bill Digest";
     final String body =
